move the counter class from nog_bdb to nog_sql
