+++ impl/nog_sql/counter.py
